setup new individuals for data set 3 and extraction or rules.
Changes to input data to accept doubles

+++ src/main/java/model/Data.java
@@ -4,23 +4,23 @@ package model;
  * Created by shaun on 18/10/16.
  */
 public class Data {
-    int[] conditions;
-    int output;
+    double[] conditions;
+    double output;
 
-    public Data(int[] conditions, int output) {
+    public Data(double[] conditions, double output) {
         this.conditions = conditions;
         this.output = output;
     }
 
-    public int[] getConditions() {
+    public double[] getConditions() {
         return conditions;
     }
 
-    public void setConditions(int[] conditions) {
+    public void setConditions(double[] conditions) {
         this.conditions = conditions;
     }
 
-    public int getOutput() {
+    public double getOutput() {
         return output;
     }
